+++ internal/privi/private_data_test.go
@@ -1,4 +1,4 @@
-package privy
+package privi
 
 import (
 	"context"
