+++ internal/privi/server.go
@@ -1,4 +1,4 @@
-package privy
+package privi
 
 import (
 	"encoding/json"
